Start migration to Chakra-UI with Provider (#2)
* Start migration to Chakra-UI with Provider

Dark Mode, Header, and Footer have functionality back as Chakra. Index
was next, and has the outermost layer of content done, but requires a
deepdive into the NFTPreview.js to reengineer for Chakra. Created a side
drawer for filters, etc for the NFT view as I'm used to seeing on
marketplaces.

* Set height on index to push footer off-screen.

Also set Footer container to be chakra.footer vice a Box